# backend/src/swagger.json
{
  "openapi": "3.0.0",
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "info": {
    "title": "Bootcamp 2025 API",
    "version": "1.0.0",
    "description": "API para gerenciamento de atividades e usuários do Bootcamp"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Servidor local"
    }
  ],
...
  "paths": {
...
    "/activities/types": {
      "get": {
        "summary": "Listar tipos de atividades",
        "tags": ["Atividades"],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "example": [
                  {
                    "id": "c7f42b7b-4b32-453c-902d-33baae755607",
                    "name": "string",
                    "description": "string",
                    "image": "string"
                  }
                ]
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "example": {
                  "error": "Autenticação necessária"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "example": {
                  "error": "Erro inesperado"
                }
              }
            }
          }
        }
      }
    },
    "/activities": {
      "get": {
        "summary": "Listar atividades com paginação",
        "tags": ["Atividades"],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Número da página",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "Tamanho da página",
            "schema": {
              "type": "integer",
              "default": 10
            }
          },
          {
            "name": "typeId",
            "in": "query",
            "description": "Filtrar atividades por tipo",
            "schema": {
              "type": "integer",
              "format": "uuid"
            }
          },
          {
            "name": "orderBy",
            "in": "query",
            "description": "Ordernar as atividades por um campo específico",
            "schema": {
              "type": "string",
              "default": "createdAt"
            }
          },
          {
            "name": "order",
            "in": "query",
            "description": "Direção da ordenação",
            "schema": {
              "type": "string",
              "default": "desc"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "example": {
                  "data": [
                    {
                      "id": "ac209fd0-6637-43f0-8fb0-99479a9d491b",
                      "title": "Yoga no parque",
                      "type": "Yoga"
                    }
                  ],
                  "meta": {
                    "totalItems": 1,
                    "itemCount": 1,
                    "itemsPerPage": 10,
                    "totalPages": 1,
                    "currentPage": 1
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "example": {
                  "error": "Autenticação necessária"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "example": {
                  "error": "Erro inesperado"
                }
              }
            }
          }
        }
      }
    },
    "/activities/all": {
      "get": {
        "summary": "Listar todas as atividades sem paginação",
        "tags": ["Atividades"],
        "parameters": [
          {
            "name": "typeId",
            "in": "query",
            "description": "Filtrar por tipo de atividade",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "orderBy",
            "in": "query",
            "description": "Campo para ordenação",
            "schema": {
              "type": "string",
              "default": "createdAt"
            }
          },
          {
            "name": "order",
            "in": "query",
            "description": "Direção da ordenação",
            "schema": {
              "type": "string",
              "default": "desc"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "example": [
                  {
                    "id": "ac209fd0-6637-43f0-8fb0-99479a9d491b",
                    "title": "Yoga no parque",
                    "type": "Yoga"
                  }
                ]
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "example": {
                  "error": "Autenticação necessária"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "example": {
                  "error": "Erro inesperado"
                }
              }
            }
          }
        }
      }
    },
    "/activities/user/creator": {
      "get": {
        "summary": "Listar atividades criadas pelo usuário (paginado)",
        "tags": ["Atividades"],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Número da página",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "Tamanho da página",
            "schema": {
              "type": "integer",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "example": {
                  "data": [
                    {
                      "id": "ac209fd0-6637-43f0-8fb0-99479a9d491b",
                      "title": "Yoga no parque",
                      "status": "pending"
                    }
                  ],
                  "meta": {
                    "totalItems": 1,
                    "itemCount": 1,
                    "itemsPerPage": 10,
                    "totalPages": 1,
                    "currentPage": 1
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "example": {
                  "error": "Autenticação necessária"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "example": {
                  "error": "Erro inesperado"
                }
              }
            }
          }
        }
      }
    },
    "/activities/user/creator/all": {
      "get": {
        "summary": "Listar todas as atividades criadas pelo usuário",
        "tags": ["Atividades"],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "example": [
                  {
                    "id": "ac209fd0-6637-43f0-8fb0-99479a9d491b",
                    "title": "Yoga no parque",
                    "status": "pending"
                  }
                ]
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "example": {
                  "error": "Autenticação necessária"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "example": {
                  "error": "Erro inesperado"
                }
              }
            }
          }
        }
      }
    },
    "/activities/user/participant": {
      "get": {
        "summary": "Listar atividades em que o usuário participa (paginado)",
        "tags": ["Atividades"],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Número da página",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "Tamanho da página",
            "schema": {
              "type": "integer",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "example": {
                  "data": [
                    {
                      "id": "ac209fd0-6637-43f0-8fb0-99479a9d491b",
                      "title": "Yoga no parque",
                      "status": "approved"
                    }
                  ],
                  "meta": {
                    "totalItems": 1,
                    "itemCount": 1,
                    "itemsPerPage": 10,
                    "totalPages": 1,
                    "currentPage": 1
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "example": {
                  "error": "Autenticação necessária"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "example": {
                  "error": "Erro inesperado"
                }
              }
            }
          }
        }
      }
    },
    "/activities/user/participant/all": {
      "get": {
        "summary": "Listar todas as atividades em que o usuário participa(precisa estar confirmado e aprovado na atividade)",
        "tags": ["Atividades"],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "example": {
                  "totalItems": 6,
                  "data": [
                    {
                      "id": "00000000-0000-0000-0000-000000000001",
                      "title": "Atividade Genérica 1",
                      "description": "Descrição genérica da atividade 1",
                      "typeId": "00000000-0000-0000-0000-000000000002",
                      "confirmationCode": "GEN12345",
                      "image": "https://link-para-imagem.com/imagem-generica.jpg",
                      "scheduledDate": "2025-04-01T07:00:00.000Z",
                      "createdAt": "2025-03-27T20:34:14.420Z",
                      "deletedAt": null,
                      "completedAt": null,
                      "private": false,
                      "creatorId": "00000000-0000-0000-0000-000000000003",
                      "type": {
                        "id": "00000000-0000-0000-0000-000000000002",
                        "name": "Tipo Genérico",
                        "description": "Descrição genérica do tipo",
                        "image": "tipo.png"
                      },
                      "ActivityAddresses": [],
                      "ActivityParticipants": [
                        {
                          "id": "00000000-0000-0000-0000-000000000004",
                          "activityId": "00000000-0000-0000-0000-000000000001",
                          "userId": "00000000-0000-0000-0000-000000000005",
                          "approved": true,
                          "confirmedAt": "2025-03-27T22:43:00.872Z",
                          "user": {
                            "id": "00000000-0000-0000-0000-000000000005",
                            "name": "Usuário Genérico",
                            "avatar": "http://localhost:3000/public/images/default-avatar.jpg"
                          }
                        },
                        {
                          "id": "00000000-0000-0000-0000-000000000006",
                          "activityId": "00000000-0000-0000-0000-000000000001",
                          "userId": "00000000-0000-0000-0000-000000000007",
                          "approved": true,
                          "confirmedAt": "2025-03-28T16:02:23.227Z",
                          "user": {
                            "id": "00000000-0000-0000-0000-000000000007",
                            "name": "Usuário Genérico 2",
                            "avatar": "http://localhost:3000/public/images/default-avatar.jpg"
                          }
                        }
                      ],
                      "creator": {
                        "id": "00000000-0000-0000-0000-000000000003",
                        "name": "Criador Genérico",
                        "avatar": "http://localhost:3000/public/images/default-avatar.jpg"
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "example": {
                  "error": "Autenticação necessária"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "example": {
                  "error": "Erro inesperado"
                }
              }
            }
          }
        }
      }
    },
...
    "/activities/new": {
      "post": {
        "summary": "Criar nova atividade",
        "tags": ["Atividades"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string" },
                  "description": { "type": "string" },
                  "typeId": { "type": "string", "format": "uuid" },
                  "scheduledDate": { "type": "string", "format": "date-time" },
                  "private": { "type": "boolean", "default": false }
                },
                "required": ["title", "typeId", "scheduledDate"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "example": {
                  "id": "ac209fd0-6637-43f0-8fb0-99479a9d491b",
                  "title": "Yoga no parque"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "example": {
                  "error": "Dados inválidos"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "example": {
                  "error": "Autenticação necessária"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "example": {
                  "error": "Erro inesperado"
                }
              }
            }
          }
        }
      }
    },
    "/activities/{id}/subscribe": {
      "post": {
        "summary": "Inscrever-se em uma atividade",
        "tags": ["Atividades"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "example": {
                  "id": "c9026925-ae55-4153-a5e6-983b878fab63",
                  "activityId": "ebb4ed62-6199-45f3-9e00-85da4f11628a",
                  "userId": "b4529157-abbb-4e36-9228-38bf5a3e0970",
                  "approved": false,
                  "confirmedAt": null
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "example": {
                  "error": "Você já está inscrito nesta atividade"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "example": {
                  "error": "Autenticação necessária"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "example": {
                  "error": "Você não pode se inscrever na própria atividade"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "example": {
                  "error": "Atividade não encontrada"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "example": {
                  "error": "Erro inesperado"
                }
              }
            }
          }
        }
      }
    },
    "/activities/{id}/update": {
      "put": {
        "summary": "Atualizar atividade",
        "tags": ["Atividades"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string" },
                  "description": { "type": "string" },
                  "typeId": { "type": "string", "format": "uuid" },
                  "adress": {
                    "type": "string",
                    "scheduledDate": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "boolean": { "type": "boolean", "format": "binary" }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "example": {
                  "message": "Atividade atualizada com sucesso"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "example": {
                  "error": "Dados inválidos"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "example": {
                  "error": "Autenticação necessária"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "example": {
                  "error": "Você não tem permissão para editar esta atividade"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "example": {
                  "error": "Atividade não encontrada"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "example": {
                  "error": "Erro inesperado"
                }
              }
            }
          }
        }
      }
    },
    "/activities/{id}/conclude": {
      "put": {
        "summary": "Concluir atividade",
        "tags": ["Atividades"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "example": {
                  "message": "Atividade concluída com sucesso"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "example": {
                  "error": "Atividade já concluída"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "example": {
                  "error": "Autenticação necessária"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "example": {
                  "error": "Você não tem permissão para concluir esta atividade"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "example": {
                  "error": "Atividade não encontrada"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "example": {
                  "error": "Erro inesperado"
                }
              }
            }
          }
        }
      }
    },
...
    "/activities/{id}/unsubscribe": {
      "delete": {
        "summary": "Cancelar inscrição",
        "tags": ["Atividades"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "example": {
                  "message": "Inscrição cancelada com sucesso"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "example": {
                  "error": "Você não está inscrito nesta atividade"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "example": {
                  "error": "Autenticação necessária"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "example": {
                  "error": "Atividade não encontrada"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "example": {
                  "error": "Erro inesperado"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}